lombok-ify and refactor code
