+++ src/main/java/dev/redcoke/mcserverping/utils/TextComponentFormatter.java
@@ -1,4 +1,4 @@
-package dev.redcoke.mcserverping;
+package dev.redcoke.mcserverping.utils;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
